requirements.txt
google-cloud-vision==2.5.0
PyMuPDF==1.18.19
Pillow==8.2.0
